feat(api): Break MiniTwit into multiple files for easy maintenance
This commit introduces the src/minitwit directory which in turn contains
sub directories for the api routes and helper modules that MiniTwit
uses, no frontend has been moved in here yet as this is mostly focused
on getting the api to work.
Big thanks to [this
stackoverflow](https://stackoverflow.com/questions/5015471/using-sinatra-for-larger-projects-via-multiple-files)
