+++ src/minitwit/api/init.rb
@@ -1,0 +1,6 @@
+require_relative 'latest'
+require_relative 'register'
+require_relative 'fllws'
+require_relative 'msgs'
+
+require_relative 'notused'
