+++ src/minitwit/helpers/format.rb
@@ -1,0 +1,15 @@
+module Formatters
+
+  # Format datetime
+  def format_datetime(timestamp)
+    Time.at(timestamp).utc.strftime('%Y-%m-%d @ %H:%M')
+  end
+
+  # Gravatar URL
+  def gravatar_url(email, size = 80)
+    hash = Digest::MD5.hexdigest(email.strip.downcase)
+    "http://www.gravatar.com/avatar/#{hash}?d=identicon&s=#{size}"
+  end
+
+end
+
