+++ src/minitwit/helpers/init.rb
@@ -1,0 +1,8 @@
+require_relative 'dbhelper'
+MiniTwit.helpers DatabaseHelper
+
+require_relative 'format'
+MiniTwit.helpers Formatters
+
+require_relative 'auth'
+MiniTwit.helpers Authorization
